GAI/requirements.txt
openai>=1.40.0
groq
streamlit>=1.37.0
python-dotenv>=1.0.1
